Fixes for inaccessible URL
