+++ .github/workflows/documentation.yml
@@ -2,15 +2,17 @@
 name: Deploy Documentation to Pages
 
 on:
-  # Runs on pushes targeting the default branch
   push:
     branches:
       - main
+    # Limit runs to only when docs change
+    paths:
+      - 'docs/**'
 
-  # Allows you to run this workflow manually from the Actions tab
+  # Allows running this workflow manually
   workflow_dispatch:
 
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+# Permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
 permissions:
   contents: read
   pages: write
